requirements.txt
#appdirs==1.4.0
#awsebcli==3.9.0
#backports.ssl-match-hostname==3.5.0.1
#BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
#blessed==1.9.5
#botocore==1.5.7
#cement==2.8.2
#colorama==0.3.7
#coloredlogs==5.2
#cssselect==0.9.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.1
djangorestframework==3.5.4
#docker-py==1.7.2
#dockerpty==0.4.1
#docopt==0.6.2
#docutils==0.13.1
#feedfinder2==0.0.1
#feedparser==5.1.3
gunicorn==19.6.0
#heroku==0.1.4
#humanfriendly==2.3.2
#jieba==0.35
#jmespath==0.9.1
lxml==3.3.5
#monotonic==1.2
newspaper==0.0.9.8
nltk==2.0.5
numpy==1.12.0
#olefile==0.44
#packaging==16.8
#pathspec==0.5.0
#Pillow==4.0.0
#pyparsing==2.1.10
#python-dateutil==2.4.0
#PyYAML==3.11
requests==2.9.1
selenium==3.0.2
#semantic-version==2.5.0
#six==1.7.3
#static3==0.7.0
#tabulate==0.7.5
#tldextract==1.5.1
#wcwidth==0.1.7
#websocket-client==0.40.0
psycopg2
pattern